pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vc-job-board"
version = "0.1.0"
description = "Aggregates job listings from 50+ VC portfolio job boards"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "playwright>=1.40",
    "pyyaml>=6.0",
    "rapidfuzz>=3.0",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
